Refactor: split ssl algorithm into serveral files
one large source files into smaller ones
- config
- dataset
- ssl algorithm
